fix: Only run other rules when required validates (#32)
* fix: Only run other rules when required validates

This changes the validator to run required first and skip all other
rules if required fails. The error from required will only be reported
if the given field is actually required. We have not found a case were
another validation should fail on an empty value without it beeing a
required error. Thus, this makes it unecessary to handle the empty
corner cases in all rules.

BREAKING CHANGE: This will result in rules not being run with empty
values so if there are rules that depends on empty values they would not
work any longer.

* fix: Remove empty value checks from rules

* fix: Alter check for if statements so it works for rules and required

* refactor: Move fetching of config for rule to a function