fix: reset errors between calls to validate
Remove the errors between subsequent calls to validate.

The validator keeps the errors from a call to validate in the instance
variable errors. This variable aggregated all errors through the
lifetime of the validator, if the error was not removed inside the
validate call.

This change makes the call to validate first remove all errors from the
previous call to the method before it runs the validations.