chore: Add babel-plugin-lodash to compress build
This reduces the size of the build by only including the parts of lodash
that is used. It reduces the size by about 100 kB, or up to 140 kB if
included in a project which already uses the same parts of lodash.

Unfortunately, if included in a project which uses lodash, but doesn't
use babel-plugin-lodash, this will increase the size by about 40 kB
because lodash will be included both as a single blob and modularly.
However, babel-plugin-lodash is easy include and I don't know of any
downsides to using it, so anyone concerned about size should include it,
which means this probably isn't a problem.